Move this out of ZooKeeper module, rework for webhook only
